backend/requirements/requirements.dev.txt
-r requirements.txt
appdirs==1.4.3
astroid==2.3.3
attrs==19.3.0
black==19.10b0
Click==7.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pathspec==0.7.0
pylint==2.4.4
pylint-django==2.0.13
pylint-plugin-utils==0.6
regex==2020.1.8
six==1.14.0
toml==0.10.0
typed-ast==1.4.1
wrapt==1.11.2
coverage==5.0.3
bandit==1.6.2
PyJWT==1.7.1
mypy==0.761
mypy-extensions==0.4.3
djangorestframework-stubs==1.1.0
django-stubs==1.4.0
